requirements.txt
asn1crypto==0.24.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.7.0
cryptography==2.1.4
cycler==0.10.0
Cython==0.29.4
dask==1.1.1
decorator==4.3.2
docopt==0.6.2
enum34==1.1.6
futures==3.2.0
h5py==2.9.0
idna==2.6
ipaddress==1.0.17
ipython==5.8.0
ipython-genutils==0.2.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
leveldb==0.194
matplotlib==2.2.3
networkx==2.2
nose==1.3.7
numpy==1.16.1
olefile==0.45.1
opencv-python==4.0.0.21
pandas==0.24.0
pathlib2==2.3.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pipreqs==0.4.9
prompt-toolkit==1.0.15
protobuf==3.0.0
ptyprocess==0.6.0
pycrypto==2.6.1
pydot==1.4.1
Pygments==2.3.1
pygobject==3.26.1
pyparsing==2.3.1
python-dateutil==2.7.5
python-gflags==3.1.2
pytz==2018.9
PyWavelets==1.0.1
pyxdg==0.25
PyYAML==3.12
requests==2.21.0
scandir==1.9.0
scikit-image==0.14.2
scikit-learn==0.20.1
scipy==1.2.0
SecretStorage==2.3.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
subprocess32==3.5.3
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
yarg==0.1.9
